Reorganize and namespace modules
